companies_act_2013/governance_db/requirements.txt
# Governance RAG System - PostgreSQL Requirements

# Database
psycopg2-binary
python-dotenv 

# Ollama API client
requests

# Flask API
flask
flask-cors

# Utilities
numpy
